requirements.txt
# Core ML
xgboost>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Technical analysis
ta>=0.11.0

# Crypto exchange / market data
ccxt>=4.0.0

# Notifications
requests>=2.31.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0

# System monitoring
psutil>=5.9.0

# Terminal output
colorama>=0.4.6
